Github Actions Python Package build and test workflow
Add a Github Actions workflow that builds and tests the academic-observatory Python package when pushes and pull requests (to master or develop) are made.

The tests are run with Python 3.7 on Ubuntu latest (19.10), Ubuntu 18.04 and macOS latest (Catalina). Extra unit tests, docstrings, missing requirements and a specific Python version have been specified.