First major round of edits to address issues with versions, imports, and other dependencies since the move of the Reporting move to AO.
This is not complete as I have gotten to a point I need to check something with Richard before proceeding, but there are a lot of changes in this commit.

I am running this with main.py for this part of the move to aid in making sure it works. This needs then to be updated to work from the command line.

There are a few imports I have commented out in this set of changes. I will tidy them up in the next round.

Defaults.py was missing in the move but had to be called in order based on the new __init__ on AO. I added back in defaults.py to AO/reports, and have edited the AO/reports/__init__ to remove things that are being called out of order. Defaults.py items could be moved into reports_util.py

AO/reports/charts/__init__ some of these are not packages so can not be imported this way. I have removed them from this file:
  chord.py
  citation_count_time_chart
  generic_time_chart
  oapc_time_chart
  output_types_time_chart

The palettes is meant to be part of defaults, but was moved into charts_utils.py. I had to move it back to defaults else other code breaks. output_types is also used in palette creation so was moved to defaults too. These need to all be untangled after the move.