Add 'is_composer' check when creating observatory home so it doesn't fail when loading dags inside composer environment.
Since the class ObservatoryConfig has the 'HOST_DEFAULT_PATH' variable which is initiated when importing the class. This variable includes the 'observatory_home()' function.

observatory_path will point to '/home/airflow/gcs/data' inside the composer environment, which is correct. However, this directory is not accessible from the airflow webserver inside the composer environment, when loading dags (and thus importing ObservatoryConfig). It is available for the airflow workers.

The added 'is_composer' check will prevent raising a FileNotFoundError. It would be better to add something like 'is_airflow_webserver', but I am not sure how to check for this.