Resolved all requested changes from the pull request.
Removed translating of geoname ids into geonames data, as discussed the geonames data will be retrieved in a separate telescope and names can be translated later with e.g. SQL statements, parsing both fundref and geonames tables.

Discovered that I was missing some earlier releases, because I was only retrieving the first page of results using the gitlab api. Changed this so that all available pages will be parsed.
Some of the earlier releases (anything before v1.0) were released on the same day, so they have the same 'released-at' date. However the month and year of actual release are in the description, which is used for their release dates.
An exception is the first release (v0.1) which doesn't describe a release date anywhere. I've set it to a release date of one month prior to v0.2.

Schema was tested with release v0.1 and corresponding date 2014-03-01.