Dynamically create environment variables in observatory.py for airflow vars/connections. Note that new vars/connections still have to be defined in the docker-compose.local.yml file.
Added 'environment' and 'data_location' as airflow variables, removed them from the original schema.

The airflow variable 'data_path' is used inside telescopes, but the value for this is hardcoded in the docker compose file. The methods that make a default config and parse the schema are constructed so that they only take in account airflow vars/connections that should be obtained from the config file.

Changed the AirflowVar and AirflowConn classes slightly so they store some more information for each object. The default values and schema specifications can be stored now.