Final updates and corrections to charts - all tests passing
A few small fixes to charts and starting work on bringing
the tests up to date. Added a test_charts_utils.py that
will generate test data using the current table definitions
and place them in the fixtures directory.

Some other small changes to tests to update the column
names to current state.