Final fixes to charting and reports and requirements.txt
Some more minor fixes to enable current group reports to
function and correct a few issues with typing of Dataframe
columns for conversion to report tables.

Reverted changes to requirements.txt with Seaborn now fixed
on 0.11.* and precipy removed because it is not strictly
required for current functionality in the observatory
platform itself.