In Telescope class, added group 'download_tasks' and 'cleanup_tasks', renamed 'before_subdag_tasks' -> 'setup_tasks'.
Added corresponding methods to add tasks to the new groups mentioned above.

Added 'add_*_task' methods to add an individual task.

Removed properties from AbstractTelescopeRelease, except the 'date_str' property. This is used in one of the implemented methods of the TelescopeRelease class, so the property is required to be present.

Changed the signature of 'make_release' method, so the only parameter is '**kwargs' and updated the 'task_callable' method accordingly.

Updated the example 'make_release' method of the OrcidRelease.