Increased cloud run memory 128 -> 256 MB
Add requirement for api key, either in query or in header. The existing keys are stored locally in a file.

Changed response, added 'returned_results', 'total_results', 'schema' and 'results' to response. Also updated the response definition in the openapi file.

Added a limit parameter, which limits the number of results returned.

Defined the query filter parameters in a list that is stored in the app context, so it can be used in the 'query' module as well.