Initial proposal for Unpaywall query from COKI Sprint #1
Following the recommendations of the Sprint #1 group the query has been
modified to use a CASE WHEN syntax for readability and has been
heavily commented.

Following the findings of the group on improved precision in the Unpaywall
data, slight modifications to the logic of the queries have been made
particularly with respect to Bronze determination.

Further additional categories will be the subject of future Sprints and
the documentation is only a first pass to investigate how best to manage
and record this.

All changes are to aggregate_unpaywall.sql.jinja2