Add archive_file resource that will trigger a new revision when the docker image for the Cloud Run backend is updated.
A text file is created when executing the observatory terraform command 'build-api-image'. The SHA256 of this file after it is zipped is used as a annotation for the Cloud Run service. The annotation is not needed for anything, but will trigger an update of the resource. Even though the resource is updated in place it is enough to create a new revision that points to the latest image.