Initial proposal for Unpaywall aggregation query from COKI Sprint #1 (#350)
* Initial proposal for Unpaywall query from COKI Sprint #1

Following the recommendations of the Sprint #1 group the query has been
modified to use a CASE WHEN syntax for readability and has been
heavily commented.

Following the findings of the group on improved precision in the Unpaywall
data, slight modifications to the logic of the queries have been made
particularly with respect to Bronze determination.

Further additional categories will be the subject of future Sprints and
the documentation is only a first pass to investigate how best to manage
and record this.

All changes are to aggregate_unpaywall.sql.jinja2

* Minor changes to commenting format and additions

Added some additional documentation and guidance at the top of the file
and changed the commenting format to use Jinja comments, with the
actual documentation text being technically correct markdown.

The shift to Jinja comments is so that the documentation doesn't get
replicated in the SQL logs.

* Updates to the query in response to Bianca's comments

No changes to the actual query structure itself but some changes
to documentation to provide greater clarity and consistency.

For comments see:
https://github.com/The-Academic-Observatory/observatory-platform/pull/350#issuecomment-775893574

* Update per pull request comments to remove extraneous FROM line